[pb] Avoid deleting root when creating a PB
When a user builds a PB they can pass an out directory to write it to.
If this directory exists already we delete it first. We do this because
some of the files in the PB may be read-only and cannot be overwritten.
These files must be deleted before we can generate another PB.

In order to help the user avoid accidents, we now prevent them from
deleting their root directory.

Change-Id: I7c0146e071fc34bb562ec9e566875e046e3f5e60
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/757865
Reviewed-by: Dave Schuyler <dschuyler@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>