[media/codecs] pad ffmpeg AVCodecContext::extradata
When the AVCodecContext::extradata field is populated, the memory
should be padded by AV_INPUT_BUFFER_PADDING_SIZE (64 bytes). This
was not done in the ffmpeg H264 codec, which caused a test to fail
with the latest version of ffmpeg. This has been fixed, as well as
similar code in mediaplayer.

Test: existing test passes with new ffmpeg

Fixed: 115036
Change-Id: I3b8a12778f798dc8249f916c35512f747ebbe24a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/758441
Fuchsia-Auto-Submit: Dale Sather <dalesat@google.com>
Reviewed-by: Stefan Bossbaly <stefanbossbaly@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>