[emu] Add KVM one-liner to ffx emu start
As described in the attached bug, the `ffx emu` tool has functionality
for determining if KVM is available for emulator acceleration on Linux
machines, but doesn't provide useful feedback when that decision is made
(when the --accel auto flag is resolved). The now-deprecated `fx emu`
tool had a verbose, user-friendly message instructing the user how to
enable KVM in the most common circumstance.

This CL adjusts the logic for detecting whether KVM is available (the
previous method was less precise) and when there is a permission-based
failure, it prints help text to the screen instructing the user, as
before.

Bug: 107985
Change-Id: Ib66e3a52adf0f96608bca5f1b941c70952a87f2e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/757731
Commit-Queue: Daniel Dougherty <doughertyda@google.com>
Reviewed-by: Victor Costan <costan@google.com>
Reviewed-by: Wayne Piekarski <waynepie@google.com>
Reviewed-by: Seth Ladd <sethladd@google.com>