[async-gunzip] Instrument Decoder
Add time measurement for the entire decode operation: reading input
buffer, decompressing gzip bytes, and reading the header and footer
values.

Additionally, this CL adds functionality for tracking the compressed
size of the bytes being inflated.

These measurements are now accessible through the ChunkStats struct and
can be used to to track the efficiency of the gunzip functionality.

ChunkStats is returned alongside bytes in the stream output, and
contains the time of the decoding operation and the compressed size of
the bytes passed into decode since the previous Stream item. This allows
callers to have more control over the output values and can track the
progression of the stats throughout the runtime of the function.

The ChunkStats struct is removed from the main `decode` API wrapper
for simplicity and can be used by the caller with no knowledge of the
stats object.

Tests were added to verify that the cumulative fetched byte size was
calculated as expected and the time of the operation was non-zero.

Fixed: 111161
Test: async-gunzip
Change-Id: Iaf7ff6fcb33255fffc2ad0075e9f6c36e1a6e002
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/738633
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>
Reviewed-by: Kevan Hollbach <khollbach@google.com>
Reviewed-by: Devon H. O'Dell <dhobsd@google.com>