[debugger] Add programmatic pretty-printing.
Hooks up a built-in function add to a child to the ExprNode tree to the
pretty-printing system.

Adds a "generic container" pretty printer for running a program that
computes the children of the FormatNode.

Adds a test for this behavior. It is not yet used in production.

Change-Id: I5eca849cec18d45b22ead9946a644fb077ee8eb1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/757237
Reviewed-by: Dangyi Liu <dangyi@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>