[mdns] fix crash in mdns service
The mdns service has been crashing trying to access the zeroth entry
in an empty list:

  instance_requestor.cc#382:
    instance_info.port_ = instance.addresses_[0].port();

The list should not be empty. While we don't have a good repro, it
appears that this is happening because services are being published
with a port number of zero. The InstanceResponder agent, which
publishes service instances, was using a non-zero port number as an
indication that instance information it retains is initialized. As
a result, when the client specifies a port number of zero for an
instance, the address list for the instance isn't populated. This
causes the undefined behavior in InstanceRequestor.

In addition to changing InstanceResponder so it handles zero port
numbers, InstanceRequestor has been changed to use an 'if' rather
than an 'FX_DCHECK' to check that the address list is not empty.
If the problem is *not* due to a zero port number, this will prevent
the mdns service from crashing and will instead log an error. If this
should occur, the mdns service may fail to discover locally-published
service instances.

Fixed: b/258110998

Test: added unit test
Change-Id: I4de758d93975866d4b5516e8b3159a1fb17d0710
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/757928
Reviewed-by: Sundarajan Srinivasan <sundarsrini@google.com>
Reviewed-by: Prashanth Swaminathan <prashanthsw@google.com>
Commit-Queue: Dale Sather <dalesat@google.com>