[ffx][fho] Make FfxMain::main take moved self
Many of the existing plugins were written with the assumption that they
get a moved value of the cmd args and porting them will be difficult
this way. This change adds some basic TryFromEnv implementations for
smart pointer-ish types like Arc, Box, Option, and Result to allow for
doing things like having an Arc<> of the Writer so tests can be written
against it, which was the motivation of taking a mutable reference
instead.

This will result in far less peppering of the code with `Clone` derives
as we migrate to the new structure.

Change-Id: I23cc67a9985a70889bba33234fb9f5e717c73332
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/758865
Fuchsia-Auto-Submit: Megan Batty <mgnb@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Andrew Davies <awdavies@google.com>