[microbenchmarks][docs] Update the docs on how to run these tests
Remove references to running "fuchsia_microbenchmarks" directly as a
shell command.  The ability to run it like that will go away when we
remove the "fuchsia_shell_package" definition from BUILD.gn.  (The
shell command is no longer exercised in Infra builds.)

Add a more up-to-date description of the different ways of running
these tests on Fuchsia.

Test: manually ran the commands listed
Change-Id: I405425313277be37900b11926db5d765ecb03e42
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/759045
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>
Commit-Queue: Mark Seaborn <mseaborn@google.com>