[kernel] Introduce ZX_RIGHT_OP_CHILDREN for VMARs
Adds a new right, ZX_RIGHT_OP_CHILDREN, that provides an additional way
to control VMAR operations and how they interact with their children.
For now the right is only implemented and defined as having any meaning
for VMARs.

Most VMAR operations are unchanged in the presence of the OP_CHIDLREN
right, as they already operated on their children, and simply gain a
restriction in the absence of the right. zx_vmar_protect did not
previously allow child operations, but with this right is allowed to
reduce permissions of mappings in its subregions.

Bug: 107058
Run-All-Tests: True
Change-Id: I2606ba4c0c0e3198838baf971d0a93624e0fa581
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/753484
API-Review: Carlos Pizano Uribe <cpu@google.com>
Reviewed-by: Carlos Pizano Uribe <cpu@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>