[input-pipeline] Move default config pkg to lib
Move the default-config-pkg package from the bin dir to the lib dir to
make it clear that this package is intended to support the library in
multiple binaries and not just the input-pipeline binary.

A follow-up CL will remove the duplicate entry in bin/input-pipeline
after out-of-tree dependencies are cleaned up.

Bug: 100664
Test: Tested manualy on device
Change-Id: If690d0b3cb0a3c6dab6d9c20b162bced746360ab
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/756183
Reviewed-by: Alexander Brusher <abrusher@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Paul Faria <paulfaria@google.com>