[sherlock][thermal] Opt disable platform metrics
Due to a change introduced in tqr/525531 crash reports were being
filed when audio was being throttled, which is expected in normal
operation on Sherlock. This allows platform_metrics_node to be omitted
from the ThermalStateHandler configs, which will disable the metrics
and crash reports for the handler. This will be used for the audio
thermal handler, while preserving thermal logs for the cpu thermal
handler, which will be added in a follow-up change.

Bug: 111311
Test: fx test power-manager-tests

Change-Id: I6d8ba349fb8437918bfae0b185d6d4d16dcdf5dd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/757263
Reviewed-by: Onath Dillinger <claridge@google.com>
Fuchsia-Auto-Submit: Kyle Gong <kgong@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>