[build][bazel] Make zx_library() generate packages for the internal SDK.
This CL ensures that a zx_library() target that is published to the
"internal" SDK category will compile properly. The issue, as explained
in the associated bug, is that by default sdk_source_set() sets the
`include_base` to `//` for the "internal" category. This allows things
like `//src/foo:bar` to be visible as `@sdk_internal//src/foo:bar`
in Bazel rules.

However, for zx_library() targets, the include base is always the
`include` directory relative to the BUILD.gn where the target is
defined (Zircon artifacts do not follow common Fuchsia conventions),
so a value of `include` should be provided for these.

Before this CL, the headers were misplaced in the resulting SDK
atom, making the sources unusable in Bazel build commands.

Bug: 114820
Change-Id: If54827a6a98fd0e21032a263c16a19eacf75506e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/760202
Reviewed-by: Chase Latta <chaselatta@google.com>
Commit-Queue: David Turner <digit@google.com>