[lowpan-ot-driver][advertising-proxy] Properly handle address changes.
In <fxbug.dev/115170>, it was noticed that we are not properly handling
cases where the host's IP addresses are changed. This CL makes sure that
we notice the case where the IP addresses for a proxied host are updated
and make sure that the appropriate changes are properly communicated to
the mDNS service.

Bug: 115170
Test: Manual Verification
Change-Id: Iaafc212a81b9bd84547ca7d614fb0be1fd05754d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/759662
Commit-Queue: Robert Quattlebaum <rquattle@google.com>
Reviewed-by: Prashanth Swaminathan <prashanthsw@google.com>