[netstack3] Apply device constraint for local delivery
When looking up routes with an egress device specified, don't return a
route with a source or destination address on a different device. This
still allows local delivery if the source and destination address are
assigned on the same device (or are identical) through the loopback
interface as long as no egress device is explicitly requested.

Bug: 112584
Change-Id: Ic96df298d9906c4a7c22e14669c7eaf05f9784e0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/759602
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>
Commit-Queue: Alex Konradi <akonradi@google.com>