[testexpects] Log mismatched expectations at end
This makes it more convenient to assess which test expectations need to
be updated from the end of a failing run.

Also adds an INFO log when expected failures are observed, so that it's
clearer from the test log that that case was expected to fail and will
be reported as a pass.

Bug: 103572
Change-Id: I08ae9e8721be8e9214a47cbe08a1cf7394c4af8c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/757964
Reviewed-by: Nick Brown <nickbrow@google.com>
Fuchsia-Auto-Submit: Evan Wildenhain <wildenhain@google.com>
Reviewed-by: Amanda Tait <atait@google.com>
Commit-Queue: Evan Wildenhain <wildenhain@google.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>