[ffx][test] Remove managed output experiment
Removes the managed output experiment and related commands from ffx
test. This feature enabled ffx test to automatically save the last few
test results to the structured output format. This feature was created
at a time when ffx test was thought to be a 1:1 replacement for tools
such as fx test, which support a full edit-compile-test workflow. Since
ffx test will not support build, as it must be build agnostic, current
thought is that it will be a tool indirectly invoked by other tools
such as bazel test or fx test, which provide the build components. In
this situation, it no longer makes sense to directly add user facing
features to ffx test. Instead, this feature might be added in higher
layers of tooling.

Bug: 113986
Testing: lsc presubmit

Change-Id: Ic0b9472b29e5db25494bf1f91b29371e19f08b98
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/752167
Reviewed-by: Ankur Mittal <anmittal@google.com>
Commit-Queue: Satsuki Ueno <satsukiu@google.com>