[wlan][tool] improve remove network usability
This change makes it easier to remove networks with the donut/ffx wlan
tools. Now networks can be removed by just providing the SSID if there
is only one network saved with that SSID. This change also adds a
message if nothing is removed after the call.

examples of arguments that should now work:
  remove-network --ssid <name>
  remove-network --ssid <name> --security-type wpa2
  remove-network --ssid <name> --credential <password>
These will now lookup an existing saved network that matches. If the
credential value is provided but the type is not specified, it will be a
Password type by default; if the type is PSK it must be specified.

Test: added unit tests, manually tested nothing removed and
combinations of security, credential not provided.

Bug: 111834
Change-Id: Ia78b8c0795da567c7d9c789fe59329bc95cbb9ef
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/756962
Reviewed-by: Hayden Nix <haydennix@google.com>
Commit-Queue: Naomi Mccracken <nmccracken@google.com>