[fuchsia-tcti] Implement a driver transport layer.
This changeset implements support in fuchsia-tcti to utilize
either an agent or driver transport for communicating with the
trusted platform module. The agent implementation is used by
general consumers to interface with the fuchsia.tpm.Transmit
protocol where as the driver implementation is intended to be
used by the cr50_agent and tpm_agent itself to implement
fuchsia.tpm.Provision and fuchsia.tpm.Deprovision. This
transport interacts directly with the driver through
/dev/class/tpm.

This change also adds an additional function to the
interface fuchsia_tpm_finalize that allows the underlying
memory to be deallocated.

Change-Id: I1db6d40f4ce46026ff84ce057be83f8368faa6c2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/758542
Reviewed-by: Aaron Green <aarongreen@google.com>
Commit-Queue: Ben Wright <benwright@google.com>