[identity] Remove last fuchsia.auth ref from account_common.
`fuchsia.auth` is a library we used to use in the prototyping days
with modular. We've been slowing removing the coupling between it and
the newer libraries in `fuchsia.identity`, this is the last thing to
remove before we can deprecate the library entirely.

Bug: 110664
Change-Id: I8ce5934eb4bb3c2a3f4e8da82aa1dfe8e98dccd2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/758736
Reviewed-by: James Buckland <jbuckland@google.com>
Commit-Queue: Jody Sankey <jsankey@google.com>