[fidlc] Fix VirtualSourceFile stack-use-after-return
This CL fixes a bug in fidlc where diagnostics would sometimes access a
SourceFile* pointing in a stack frame after it had returned.

Diagnostics hold pointers to source files, so it makes sense that source
files must remain in scope while printing diagnostics. But there is a
special VirtualSourceFile used for things synthesized by the compiler.
This belonged to the Libraries object, which went out of scope. The fix
was to explicitly pass in a VirtualSourceFile and keep that in scope.

It's still a bug if diagnostics are referring to these files, but now we
will see the diagnostic followed by a panic, not an ASan failure.

Test: manually ran fidlc on patch given in fxbug.dev/115341
Fixed: 115341
Bug: 8027
Change-Id: I2027b82db81edd63ce2598cd058d2325ba4dd2a7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/759828
Reviewed-by: Tamir Duberstein <tamird@google.com>
Fuchsia-Auto-Submit: Mitchell Kember <mkember@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>