[forensics] Sync SnapshotPersistence with disk
It's possible for the filesystem to delete a snapshot independently of
Feedback. Because calls to SnapshotPersistence::Contains or
SnapshotPersistence::SnapshotLocation are required before safely
proceeding with any operation in SnapshotPersistence, said functions
provide a natural point in time for us to sync our in-memory knowledge
of the filesystem with the true current state of the filesystem.

Bug: 102479

TESTED='fx test //src/developer/forensics'

Change-Id: I62238504f8a83f748cfaebb6fd88dc6c85101ca8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/758882
Reviewed-by: Alex Pankhurst <pankhurst@google.com>
Fuchsia-Auto-Submit: Thomas Gales <tgales@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>