[check-licenses] Updates to project filtering logic.
Adjusting the way that GN targets are filtered in the util package.
Moved the project filtering logic from the cmd package to the project
package.
With this new structure, it's easier to map GN targets to Project
objects.

Also added tests to the util package, to verify it is working properly.

Change-Id: I83409899d5692a251b718e8cda9b9bcd7b8ad310
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/758738
Reviewed-by: Yuxiao Sun <yuxiaosun@google.com>
Fuchsia-Auto-Submit: Jerry Belton <jcecil@google.com>
Commit-Queue: Jerry Belton <jcecil@google.com>