[build] Drop test environment `is_emu` field
Reasons:
* It is rendundant since this information is easily divined from the
provided dimensions in already well-established ways.
* It represents a false degree of freedom; its value is trivially and
uniquely determined by `device_type`.
* `device_type` is already part of the "platform-infra" contract (for
today at least): these represent either Swarming bot dimensions or
emulator strings that need to be introspected upon in order to perform
its most basic duties - so there is little need, effect, or precedent
for avoiding equating "is emulator" with "is 'QEMU' or 'AEMU'". That
said, this change doesn't actually figure into Infra and its
dependencies: we stay within the platform and move the "is emulator"
synthesis from the producer of tests.json to the golang/tooling
source-of-truth for its schema (//tools/build) as a convenience method,
a site that by definition cannot drift out of sync from the metadata
pipeline.
* Emulators are not something that need generalization and abstraction
of this sort; we have two - and should follow the folk wisdom of
"one, two, many" and not prematurely optimize for "many" - and this will
likely just soon be one.

Bug: 113961
Change-Id: I0224d029fad5707d34f0e02a20521e63af8120dc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/757982
Reviewed-by: Ina Huh <ihuh@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>