[netstack3] Avoid panic when v4 and v6 TCP sockes co-exist
This happened on workstation builds because we used an v4 listener id
to register for a v6 listener. This CL tags all socket ids with the
IP version so that we are less prone to such problems in the future.

Fixed: 112604

Change-Id: I6eef6656e864f4ecd3092c94211f1d5220082fd9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/761049
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Fuchsia-Auto-Submit: Zeling Feng <zeling@google.com>