[fidlc] Add unknown interaction migration mode.
Add a new experiment to fidlc which enables migration mode for unknown
interactions. This allows unknown interactions to be used but leaves the
defaults for protocols and methods with no modifiers set to
closed and strict respectively, which maintains backwards compatibility
with existing FIDL files.

This mode will be enabled first during the migration so that modifiers
can be added without breaking compatibility.

Bug: 88366
Test: fx test fidlc-test
Change-Id: I64539619ada0052307ba7cb89d2c977c6e1add7a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/759882
Reviewed-by: Mitchell Kember <mkember@google.com>
Commit-Queue: Zachary Stewart <zstewart@google.com>