[pw_auth] Change method signature of handling request streams.
Each `PasswordInteractionHandler` instance can only ever handle a
single request stream, the hanging get dirtiness and state machine would
be invalid if we gave it a second stream.

Previously the method signature didn't enforce this, here it does
because we consume self when handling a request stream. In addition we
move the stream onto the request_stream method for simplicity and
similarity with other FIDL modules, and remove the `Rc` which wasn't
needed.

Change-Id: I01db08a1acfd265b51378834b6b72c9f81cd02b1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/757926
Reviewed-by: Olivia Nordquist <nolivia@google.com>
Commit-Queue: Jody Sankey <jsankey@google.com>