[zx] Merge zxc into zx
When compiling for Fuchsia userspace, the full set of sources are
exposed.

In other cases (such as from kernel or from host), only the `zx::result`
type is available.

This also publishes zx::result to the partner SDK.

The fit::result part of the zxc README is moved into fit.
The zx::result part of the zxc README is moved into zx.

Forwarding references are left to do a soft migration.

RELNOTES: If you were developing out-of-tree and depending on the zxc
library, now you should depend on the zx library from the SDK.

Bug: 115471

Change-Id: I5fc78a3e7f5b5b361e305a538ae3d517ab7f3679
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/760468
Commit-Queue: Yifei Teng <yifeit@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>
API-Review: Adam Barth <abarth@google.com>
Reviewed-by: Adam Barth <abarth@google.com>