[fidl][cpp] method_ordinal -> event_ordinal
Change method_ordinal to event_ordinal in UnknownEventMetadata. This was
an oversight when the UnknownEventMetadata type was originally
copy-pasted from the UnknownMethodMetadata type (which was originally
called UnknownInteractionMetadata and considered for use in both methods
and events). This produces clearer, more consistent naming. I choose to
leave the name as method_ordinal/event_ordinal unlike Dart and Rust
where the name is just `ordinal` because C++ has less of a clear
standard for documentation comments than either of those languages, so
it's more useful to ensure that the name conveys that this is a
method/event ordinal as opposed to any of the other uses of the term
'ordinal' within FIDL.

Bug: 88366
Change-Id: I1dcad116a47a77ca3fb71a61a511732b2a609a79
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/761623
Commit-Queue: Zachary Stewart <zstewart@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>