[pbm] Refactor pb_get to always add the repository
The attached bug explains how the ffx tools lack an easy way to add a
repository based on a product bundle URL if it already exists in the
local cache. Normally, the repository is set up during `product-bundle
get`, but if the repository is removed, it's much harder to get it back.

This CL splits the pb_get function into functional sections, and then
tweaks the behavior so that an already-existing bundle will still have
the repository add call.

It also removes the locally-built (in-tree) bundle from the bundle list
during a get call, so that the selection code will never attempt to
download or setup a repo for the local bundle.

Bug: 113156
Change-Id: Ic5cc11e78d34715af09857d4bbe3f3f25eb827ac
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/757722
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: Dave Schuyler <dschuyler@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>