[superproject] Roll third_party/pigweed pw_toolchain: Add host_clang coverage toolchain
Adds a coverage toolchain, similar to the existing asan, msan, tsan,
etc toolchains, that allows creation of concrete toolchains that does
not depend on the state of `pw_toolchain_SANITIZERS` to enable coverage
instrumentation.

This came about in a downstream project when it was hard to make a
coverage toolchain reusing the logic from `host_debug` since
`pw_toolchain_SANITIZERS` could not be set until after the
`pw_toolchain_host_clang` scope would have been interpreted. Instead of
doing this logic in the downstream project, we can do it in Pigweed,
which makes it easier to ensure it will be properly updated with the
others if the pattern changes.

Original-Bug: b/254891135
Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/119633
Original-Revision: ed22dd3c8f8ba79cf7d55a8d0bcca09c3b5c5d53
GitOrigin-RevId: 113ded054cd7670f26f9d6db849ec8173e0b4019
Roller-URL: https://ci.chromium.org/b/8797220475035822913
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I5bb13a654705ef145f88823750f76b0509a218f1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/761883