[assembly] Rename Image Assembly Config targets
In the assembled_system.gni template, we have multiple references to a
file known as the image assembly config. One of these files is used as
an input to the creation of the legacy bundle, and the other is an
output of the Product Assembly tool. To clarify the usage of each file
and align with other canonical terms in Software Assembly, these labels
and file references need to be updated in the assembled system template.

Dependencies: N/A

Tests: `fx set core.x64 && fx build`
Change-Id: I7c591bfd12c0c1f0e3d454aca85a096ed62a801d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/760023
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>
Commit-Queue: Daniel Young <youngda@google.com>