[audio][drivers][da7219] Fix server management in dfv2
Since the framework requires allowing a connection in addition
to the connection from a client, we allow multiple bindings of the
server, but share the instance.

We still limit the support to one server (and hence one client) in
ServerConnector, but allow re-connecting after an unbind.

Bug: 102079
Bug: 102096

Test: runtests -n device-enumeration-test on Atlas
using the chromebook-x64-dfv2 board.

Change-Id: I77eb4d8739c13058d4a4500893c97333debfbbf6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/760982
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Sarah Chan <spqchan@google.com>
Commit-Queue: Andres Oportus <andresoportus@google.com>