[libc][musl] Fix data segment partial page fill
Fix an issue with the partial zero'ing of a data segment's partial
page fill during `map_library`.

Background:

When MUSL maps a DSO into a process, it receives a pager-backed VMO
representing the .so file and all of its contents.  Somewhere in this
file, there will (typically) be a .data segment image, sometimes
followed by a .bss image.

The ELF file tells the loader where in the file this data segment
image is, as well as where it need to be loaded in the virtual address
space.  It also tells the loader how many bytes the segment is in the
file, as well as how many bytes in RAM the segment should take (file
bytes <= RAM bytes).  This allows the file to append the .bss
immediately after the data segment by simply specifying an in-RAM size
which is larger than the in-file size.

So, the loader creates a COW-clone child of the portion of the
pager-backed VMO which contains the initial data segment image, and
then grows this VMO with any extra pages required in the case that
in-RAM size > in-file size.  If/when the program attempts to access
any of the extra pages, they will be supplied to the mapping on demand
by the VM with the zero-page (first read fault), or PMM allocated
pages which have been zero'ed (first write fault).

This is not quite enough, however.  The data segment image which
exists in the file does not _have_ to fill out to the end of page
boundary.  Instead, the ELF file is allowed to have whatever it wants
immediately after the end of the data segment.  Because of this, the
loader must fill the out the final partial page of the data segment
with zeros in the case that the segment does not end on a page
boundary.  Writing zeros to this last partial page will immediately
trigger the COW-clone behavior, but it will result in a proper
in-memory representation of the data segment for the DSO.

The Issue:

Unfortunately, there was a small bug in this code.  Instead of filling
out the final partial page, the loader was zeroing out until the end
of the VMO itself, even if it had been extended to cover a .bss which
took even more pages.  This causes the pages to be immediately write
faulted in (fetching new pages from the PMM) where they (currently)
remain resident for the life of the program (unless they happen to be
harvested by the zero page de-duper)

Typically, this is a difficult thing to notice.  Technically, the
image loaded is valid, it just happened to fetch more pages at startup
then it normally would have needed to.  It became a bit harder to
ignore, however, when loading `libclang_rt.asan.so`.  This DSO, loaded
only during ASAN builds (but by every program) only has a 0x180 byte
data segment, but it has 2514 pages of bss following this data
segment.  The amount of page faulting, and subsequent VMM lock
contention, which occurs when multiple programs are started at the
same time was impacting some stress test timing to the point that it
became hard to not notice that something was off.

The Fix:

Simple.  Just don't do that.  Only zero-fill out to the end of the
final partial page, not until the end of the mapped VMO.

Change-Id: I3d2f15a4290add9cf3a28660f7b61b8afc606d40
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/762184
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: John Grossman <johngro@google.com>