[ermine] Display loading indicator when app is launched
This change displays a spinning loading indicator next to the
application name in the app launcher UI. The indicator is shown
immediately after the entry is tapped or selected. It is hidden after
the view for the application is created.

Note: This spinning indicator is in addition to the loading indicator
that is displayed when the underlying component is started, but before its
view is created.

Bug: 113725
Test: Updated unittests pass.

Change-Id: I56d7567ee55717a7d5d58b81a0371dc22592b58e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/762402
Fuchsia-Auto-Submit: Sanjay Chouksey <sanjayc@google.com>
Reviewed-by: Filip Filmar <fmil@google.com>
Commit-Queue: Sanjay Chouksey <sanjayc@google.com>