[icu] Pull in references to icu_root under //src/lib/icu/*
With the idea to limit the visibility of ICU libraries in
Fuchsia to the tree under `//src/lib/icu`.  This makes
referencing ICU easier for the user, and allows us
some leeway to control access to ICU so that we can
later implement ICU version choice for the product
assembly process.

Bug: 114723
Change-Id: I4f966b6a7f0a7c703277d16aa563a57bb79ab2ef
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/762327
Reviewed-by: Konstantin Pozin <kpozin@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>