[fidl] Migrate some C++ FIDL bits users to a simpler bit clearing API
`a - b` will return all bits present in `a` and absent in `b`.

It's preferred over `a & ~b` not just because it's simpler, but also
because `a & ~b` will also implicitly drop any unknown bits in `a`,
whether the user intended or not. The dropping happens because `~b`
only returns known bits that are absent in `b`, as opposed to any bit
representable in the underlying type that are absent in `b`.

Change-Id: I473e29924138cba3ec6ac651e03a96a7aa6d78e7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/760822
Commit-Queue: Yifei Teng <yifeit@google.com>
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>