[fuchsia-hyper-test-support] Support running on host
This allows fuchsia-hyper-test-support to run on host, using the
async-net library to implement async networking. It unfortunately
has slightly different semantics than fuchsia_async::net, so we
need to implement a few OS-specific wrapper to get everything to
compile.

In order to implement this, we need to allow tokio-rustls to
be used on both host and target.

Change-Id: I31f2b6aea149ceaee6f2d4b221df38d523253f82
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/753809
Reviewed-by: Sen Jiang <senj@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Dave Schuyler <dschuyler@google.com>
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>