[fxfs] Change locks to be fair and more efficient
This changes LockManager so that locks are taken fairly and when locks
are released, only the necessary tasks are woken rather than all of them
as was previously the case.

This addresses the flake where too many tasks got woken up which caused
the executor's port queue to overflow.

Fixed: 113879
Multiply: fxfs-tests
Change-Id: I252d800bed6ac446aaddcdf9aafa13081d514f32
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/754424
Reviewed-by: Chris Tam <godtamit@google.com>
Commit-Queue: Chris Suter <csuter@google.com>